requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
certifi==2022.9.24
charset-normalizer==2.1.1
Django==4.1.2
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-api-key==2.2.0
filelock==3.8.0
idna==3.4
lxml==4.9.1
PySocks==1.7.1
python-decouple==3.6
pytz==2022.4
requests==2.28.1
snscrape==0.4.3.20220106
soupsieve==2.3.2.post1
sqlparse==0.4.3